requirements.txt
asgiref==3.3.1
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.7
django-js-asset==1.2.2
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
futures==3.1.1
google==3.0.0
googletrans==3.1.0a0
goslate==1.5.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pillow==8.1.2
pycodestyle==2.7.0
PyDictionary==2.0.1
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
sniffio==1.2.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
Wikipedia-API==0.5.4
